Set `VersionHistory` to `init` only
With mapped collections, items are added to the collection after they are mapped. With compatible properties, however, the property value is simply set as a reference to the original object. If the property is a collection, there's no evaluation or mapping of the individual items in the collection. As a result, the property must be settable—even though it's a collection, and we don't normally want collection-valued properties to be settable.

Previously, that necessitated the suppression of `CA2227`. With the introduction of `init`, however, we can replace the `set` with `init` to workaround this. Once we target .NET 5, this won't work and we'll need to reintroduce the suppression. But even then, `init` better communicates the expectations for this property. And, in fact, most new view models are actually implemented as `record` types to enforce this.