Updated tests to use .NET 5
Currently, none of the core OnTopic libraries rely on .NET 5 specific capabilities and, therefore, none of them multi-target .NET 5. That said, we _expect_ most implementors to be using .NET 5 and, therefore, it makes sense that the test hosts will be running in .NET 5 themselves. That should also provide them with some minor performance benefits.

For reasons that aren't entirely clear, migrating the `OnTopic.Tests` project to .NET 5 required introducing its own copy of `IsExternalInit`. This is surprising since a) this is available in `OnTopic`, which makes its internals visible to `OnTopic.Tests`, and b) this class is available in .NET 5. Regardless, an easy fix and one we don't expect to affect most scenarios, since internals aren't usually made available outside of unit tests.