Updated `Host` to use .NET 5
Currently, none of the core OnTopic libraries rely on .NET 5 specific capabilities and, therefore, none of them multi-target .NET 5. That said, we _expect_ most implementors to be using .NET 5 and, therefore, it makes sense that the host example will be running in .NET 5 itself. That should also provide it with some performance benefits.