Reverted `ModelType` property to be `virtual`
While the recent update to avoid read-only `virtual` properties (532c49d) generally makes sense, and should be maintained for most types, it doesn't make sense for `ModelType` as some implementations of this actually need to be dynamic, as they conditionally rely on the _current_ value from `Attributes`. As the `Attributes` aren't populated until after the constructor has initialized the object, that results in the `ModelType` never being updated based on these values. Oops!