Add `SourceLink` to each of project that will be distributed via NuGet
This ensures that the symbols published as part of the symbols package (ff7833a) can be tied back to specific commits of specific source files on GitHub, and thus should allow implementors to step into OnTopic code when debugging their applications. This is a build-only dependency, and won't introduce additional dependencies for implementors.