Moved `[Obsolete]` types into new `Obsolete` folder
The `Obsolete` folder is _not_ reflected in the namespaces; each type retains its _original_ namespace, so that the obsolete message will be displayed based on historic references. The `Obsolete` folder is _exclusively_ intended for organizational purposes, to help keep the `[Obsolete]` types out of the way of the production types, so that they don't get confused with the production types, or introduce clutter while working on the product.