Reintroduced `GetTopic()`, indexer as `[Obsolete]`
The `TopicCollection` and `ReadOnlyTopicCollection` were renamed to `KeyedTopicCollection` and `ReadOnlyKeyedTopicCollection`. In their place, a non-keyed version was introduced. Callers of these classes may still expected `GetTopic()` to be in place. We obviously can't mark `TopicCollection` and `ReadOnlyTopicCollection` as `[Obsolete]` since they've been reintroduced in a new capacity, but we can introduce these common members as `[Obsolete]` as a means of instructing callers how to migrate to the new classes.