Reintroduced events as `[Obsolete]`
The `DeleteEvent`, `MoveEvent`, and `RenameEvent` were renamed to, respectively, `TopicDeleted`, `TopicMoved, and `TopicRenamed`. This provides instructions to callers of the original class name.

Unfortunately, an `[Obsolete[]` member of an interface still needs to be implemented on each implementation of that interface. As such, we won't want to persist these for long after the migration to OnTopic 5.0.0. That said, fortunately, we expect all implementations of `ITopicRepository` to implement `ObservableTopicRepository` as a base class and, therefore, can implement these in that one location in order to cover all derived types (e.g., `SqlTopicRepository`, `CachedTopicRepository`, &c.)