Reintroduced `Save()` overload as `[Obsolete]`
The `Save(topic, isRecursive, isDraft)` signature was updated to remove the `isDraft` parameter. By reintroducing the original overload, we provide instructions to callers of the original method signature. In practice, we don't expect this to be a common scenario, as the `isDraft` functionality was never fully implemented, and is an artifact of OnTopic 2.x.

Unfortunately, an `[Obsolete[]` member of an interface still needs to be implemented on each implementation of that interface. As such, we won't want to persist these for long after the migration to OnTopic 5.0.0. That said, fortunately, we expect all implementations of `ITopicRepository` to implement `ObservableTopicRepository` as a base class and, therefore, can implement these in that one location in order to cover all derived types (e.g., `SqlTopicRepository`, `CachedTopicRepository`, &c.)

As this introduces ambiguities in the XML Doc references, those also needed to be updated to refer explicitly to the signature of the non-obsolete overload. That's probably a best practice anyway, so will be fine even after we remove the obsolete overload.