Reintroduced `FromList` method as `[Obsolete]`
This was removed from `ReadOnlyKeyedTopicCollection<T>` since the use case is better covered by the similar constructor overload, but should have been kept with an obsolete notice to provide instructions to callers.

In addition, the `ReadOnlyKeyedTopicCollection<T>` used to be named `ReadOnlyTopicCollection<T>`, from which a `ReadOnlyTopicCollection` was derived. The `ReadOnlyTopicCollection` no longer derives from `ReadOnlyKeyedTopicCollection<T>`. But since callers may confuse it with the previous class of the same name—which was based on a `KeyedCollection`—I've introduced it there as well, in order to offer instruction to implementors.