Throw exception on `[FilterByAttribute("ContentType", …)]`
The `ContentType` is no longer stored as an attribute, so models attempting to `[FilterByAttribute("ContentType", …)]` will never return results—or, perhaps worse, operate off of potentially stale results from prior to the migration, yielding inconsistent results. To help avoid this situation, this will throw a runtime exception when a user attempts to map a model with this setting.

A code analyzer would be a more sophisticated approach, since this is fundamentally a design-time problem. This is a quick and easy alternative that should prevent this scenario from being introduced, while giving clear guidance to implementations that have already implemented this from previous versions.