Account for variability in `DerivedTopic` attribute descriptor name
The (legacy) `DerivedTopic` attribute descriptor varied in name. In most databases, it is `InheritedTopic`, but in some newer databases it is `TopicId` or `DerivedTopic`. The variability of the attribute key is permitted by a legacy artifact of the fact that, originally, the `TopicReferenceAttribute` had a hard-coded value of `TopicId` and, therefore, it didn't matter what the key was. As this was later extended to support general topic references, and not _just_ base topic references, this was modified, and thus implementations of e.g. OnTopic 4.6 use the `TopicId` key.

It's worth noting that one of the objectives of the `BaseTopic` handling in OnTopic 5.0.0 is to apply uniform nomenclature around this concept, with unified attribute descriptor, attribute keys, etc. This helps deliver on that objective.