Alter new `[dbo].[Topics]` columns to be `NOT NULL`
These columns are initially added as nullable since they don't have a default value. But the script will immediately populate those values, and the result should remove any remaining null values. Once that's done, we can change them to not be nullable, as the final schema expects.

This will not only validate that the previous query functioned properly, but also avoid issues with the auto-generated schema comparison script, since it isn't able to update individual columns if there are existing records.

While I was at it, I cleaned up the formatting of the original script for adding these columns.