Resolved `cref` references within XML Docs
Not all `<see cref />` references resolved. Sometimes this was due to references to old or incorrect identifiers, but more often simply out-of-scope references. In many cases, these can be resolved by simply adding a `using` statement. In other cases, it required updating the references—and, especially, the method signatures. This resolves CS1574.