Fixed bug in `IsSelected()` implementation
The `IsSelected()` implementation still suffixed the `webPath` and `WebPath` with a `:`, which was needed with `UniqueKey` to ensure it wasn't a partial match against an individual key (e.g., we didn't want `Root:Web:Pro` to match `Root:Web:Products`). With `WebPath`, that shouldn't be necessary, since it always ends in `/`.

That said, as web servers can be forgiving in trailing slashes, we still append a `/` at the end of `webPath` to be safe; this may result in two `//` at the end, but that won't impact the logic.