Abstract navigation components from `INavigationTopicViewModel`
The `INavigationTopicViewModel` defines properties that are expected to be used by views. The `NavigationTopicViewComponentBase<T>`, and its derivatives, don't need to know about most of that information. Instead, they simply need to know about the `IHiearchicalTopicViewModel<T>`. As such, by relying on this interface instead, we offer more flexibility to navigation component implementations over what types of view they return.

That said, in practice, we anticipate that _most_ implementors _will_ want to implement the members associated with `INavigationTopicViewModel`, and so we continue to support that interface as a convenience for standardizing how navigation view models are implemented, and especially with regards to determining if the current node is currently selected (via the `IsSelected()` method).