Collapsed `OnTopic.ViewModels.Items` back into `OnTopic.ViewModels`
Originally, `OnTopic.ViewModels.Items` was in the `OnTopic.ViewModels` namespace. To help organize the growing library, I moved the item types into the new `Items` folder—and, with that, namespace. But these don't really make sense in their own namespace, even if it makes sense organizationally. Given that, I've renamed the folder from `Items` to `_items` to help differentiate it from namespaces, and moved each of the items back into the `OnTopic.ViewModels` namespace.