Collapsed `OnTopic.ViewModels.Collections` back into `OnTopic.ViewModels`
Originally, `OnTopic.ViewModels.Collections` was in the `OnTopic.ViewModels` namespace. To help organize the growing library, I moved the `TopicViewModelCollection` into the new `Collections` folder—and, with that, namespace. But this doesn't really make sense in its own namespace, even if it makes sense organizationally. Given that, I've renamed the folder from `Collections` to `_collections` to help differentiate it from namespaces, and moved the `TopicViewModelCollection` type back into the `OnTopic.ViewModels` namespace.