Fixed assignment of `CurrentWebPath`
In a previous update, we replaced `NavigationViewModel<T>`'s `CurrentKey` property with `CurrentWebPath`, in anticipation of it operating off of the URL, instead of the `UniqueKey` (8fd4d80f). Unfortunately, in doing so, I missed that the _assignment_ of the `CurrentWebPath` was still pulling data from `CurrentTopic.GetUniqueKey()`—and, thus, calls to e.g. `NavigationTopicViewModel.IsSelected()` were failing, since it is now comparing against `webPath`.