Introduced new `IKeyedTopicViewModel` interface
One of the places where we rely on `ITopicViewModel` is for the `KeyedTopicViewModelCollection`. But that only necessitates a `Key`, so that each instance can be indexed by the collection. Obviously, the generic types will almost certainly have more properties, but as those are strongly typed once the generic has been constructed, that's not an issue. Given that, requiring the (overly) comprehensive `ITopicViewModel` is excessive. This provides a (very) lightweight, base interface for keyed (view) models.