Mark `IPageTopicViewModel` as obsolete
There's no expected use case for the `IPageTopicViewModel`. We foresee the limited but potential need for e.g. shared view models that will handle navigation, but these can be handled via `INavigableTopicViewModel` and `INavigationTopicViewModel<T>`. Others _may_ need additional details, which are available via `ITopicViewModel`. But we don't expect reusable _layouts_, which is where the additional requirements of e.g. `IPageTopicViewModel` comes in.

On an assessment of OnTopic 4.x implementations, it doesn't appear anyone is utilizing `IPageTopicViewModel`, except possibly on model definitions, so it should be safe to remove. Marking it as obsolete.