Exclude `Host` types from code coverage
The `Host` project is intended to aid in integration testing, and should not be subjected to testing itself.

Note: This could be excluded entirely as part of e.g. a `runsettings` file. That introduces headache with our continuous integration settings, however, since we centralize the task group that includes the unit tests, and not all projects using that task group currently have or need a `runsettings` file.