Exclude `TestDoubles` types from code coverage
The `TestDoubles` project is intended to aid in integration testing, and should not be subjected to testing itself.

In practice, most of these will end up being covered by unit tests anyway, given the nature of purpose of them. But as some types and members may be for use with external tests, not all will be used.

Note: This could be excluded entirely as part of e.g. a `runsettings` file. That introduces headache with our continuous integration settings, however, since we centralize the task group that includes the unit tests, and not all projects using that task group currently have or need a `runsettings` file.