Exclude unit tests from code coverage
Unit tests do not need to be subjected to unit testing.

In practice, as unit tests should only include the setup necessary to satisfy their assertions, they usually result in 100% coverage anyway. But there are still false positives in some scenarios. And, beyond that, having them counted as part of code coverage artificially increases the coverage percentage since unit tests introduce a lot of code which will be seen as covered.

Note: This could be excluded entirely as part of e.g. a `runsettings` file. That introduces headache with our continuous integration settings, however, since we centralize the task group that includes the unit tests, and not all projects using that task group currently have or need a `runsettings` file.