Exclude testing types from code coverage
The `OnTopic.Tests` project contains a number of types intended exclusively for aiding in unit tests, and should not be subject to unit testing themselves.

In practice, most of these will end up being covered by unit tests anyway, given the nature of purpose of them. But some types include additional logic or defensive code to provide flexibility in case they are needed in other tests in the future. In addition, there tend to be some false positives in the code coverage analysis.

Note: This could be excluded entirely as part of e.g. a `runsettings` file. That introduces headache with our continuous integration settings, however, since we centralize the task group that includes the unit tests, and not all projects using that task group currently have or need a `runsettings` file.