Address dereference of a possibly null reference
This can typically be done in unit tests by simply using the null-conditional operator, and marking the `Assert` to accept a nullable type. When warning for code outside of an `Assert`, this can usually be resolved with either a `Contract.Assume()` guard clause, or a null-forgiving operator. This resolves CS8602.