Removed unnecessary suppressions
Ideally, we should avoid the `<NoWarn />` element as while it's recognized by Visual Studio, it's not recognized by e.g. Azure DevOps, and thus we end up with a lot of false positive warnings in our CI/CD pipeline. We've instead addressed most of these inline. The two exceptions are the CS1591 (which we'll leave in place) and the CS1707 (which we'll address in a subsequent commit).