Suppress CA1707 globally
Typically, public members should not have underscores in them. That's guidance Ignia adheres to. But popular conventions use underscores in unit test names. Given that, we're suppressing this warning for the unit test project.

Note: This used to be suppressed in the `csproj` via the `<NoWarn />` element. That has the downside of not being recognized by e.g. Azure DevOps CI/CD pipelines (e0facec2), however, so we're moving it to an assembly-level suppression.