Removed unnecessary suppressions (cont.)
The CA1024 is, apparently, no longer considered necessary in this case. I'm not entirely sure why, but since the suppression was only put in place to satisfy code analysis, and code analysis no longer wants it, this should be removed.

This was missed in a previous update that covered a similar issue, as code analysis hadn't yet identified the redundancy at that point (be2d25fa).