Introduced test for alternate `NavigationRoot`
Topics can optionally configure an alternate `NavigationRoot` from which they wish the navigation to be displayed. If configured, this should be used instead of searching for the navigation root in the current context. In this test, I extended the `StubTopicRepository` to include a new page under `Web_3` which defines a `NavigationRoot`, and then confirm that navigation root is correctly delivered when that page is set to the current context.