Updated `TopicControllerTest` to use the new `TestTopicRepository`
The `TestTopicRepository` provides a more purpose-built set of use cases for evaluating the `TopicController`, `RedirectController`, and `SitemapController` (5ebc4b0a), and thus allows us to not only simplify the unit tests—and especially for the `SitemapController`—but also makes them more readable by using e.g. key values that relate directly to the feature being tested.