Removed inline definition of `ActionContext`
With the establishment of a purpose-built `TestTopicRepository` (5ebc4b0a), there is no longer a need to establish a separate `ActionContext` for the `SitemapController`. (Though, actually, this may not have been necessary previously, either, now that I look at it?) Regardless, it is unnecessary and can be removed, in preference for the global `_context` variable.