Update attribute test to use `Extended` action
In practice, this is the same as it was previously, since the `Extended()` action is a passthrough to the `Index(false)` action. Nevertheless, this makes it more explicit, extends the test coverage, and yields a more intuitive test name, since evaluating excluded attributes _only_ makes sense in context of the extended sitemap.