Introduced unit test for included sitemap attributes
We already have a unit test for ensuring that _excluded_ attributes aren't included in the `SitemapController` (561a227d), but we didn't have one for verifying that _included_ attributes are properly retained.

Technically, this could all be done via one test, but breaking it up helps yield more targeted units, in order to focus potential failures down to one specific set of functionality.