Establish test for validating that invalid keys don't throw exceptions
Routes will include URLs entered by users (or bots) which may not map to valid topic keys. In those cases, the underlying `KeyedTopicCollection.GetValue()` will throw an exception when attempting to load that key. That's usually a desirable outcome. In this case, however, we want invalid URLs to return `null`—just as they would if they were valid keys that didn't exist—thus resulting in a 404, not an exception. This functionality is already in place, but wasn't validated by a unit test.