Introduced test to evaluate `TopicMappingService` caching
Caching in the `TopicMappingService` allows the multiple associations that point to the same topic to return the same view model, thus preventing the need for that topic to be remapped. In this sample case, one association is mapped, while the other is not, thus validating that only associations specified by `[Include()]` are mapped.