Introduced test to evaluate `TopicMappingService` progressive caching
Progressive caching allows for a cached model to be extended based on the source configuration by filling in any associations needed by the subsequent request, but which may not have been needed by the original request. Since this is still using the original cached model, those associations will be available in each case where the cached model is referenced, even though each only needs a portion of the association.