Renamed to `TopicAssociationsViewModel`
The `RedundantItemTopicViewModel` identifier was confusing since a) it implied there was something significant about the two properties using the same type (there is, but not for the obvious reasons), and b) closely tied it to the `RedundantTopicViewModel` (which it is tied to—but not exclusively). The name `TopicAssociationsViewModel` offers it a bit more flexibility in how it's used, while also being more descriptive about what it contains.