Introduced test to ensure "undeleted" attributes undo `IsDirty()`
This is a bit of an outside use case, but we have logic in place so that if you delete an attribute and then recreate it, it is removed from the `DeletedItems` collection. As a result, if you recreate the attribute as `!IsDirty`, and there are no other dirty attributes, that should effectively reset the `IsDirty()` state of the attribute collection to `false`. This helps prevent topics from being marked dirty just because they were modified, if the modification had no actual impact.

In practice, this is a rare scenario as it requires for callers to explicitly set `IsDirty`. One scenario where this occurs, however, is when OnTopic is operating against `AttributeRecord` instances directly. E.g., `MarkClean()` copies an `AttributeRecord` to update its `IsDirty` flag to `false`, then removes the original `AttributeRecord` and inserts the updated copy, In that scenario, the attribute will be added and then removed from `DeletedItems`, thus resulting in that attribute being clean in the end.