Introduced test for generic exception types with unexpected constructors
When calling `Contract.Require<T>()` with a custom exception type `T` and a `errorMessage`, we expect that there will be an overload of the `T` constructor that accepts a single string value for the error message, as is a convention in .NET. While this is a _popular_ convention and a best practice, however, it is not _enforced_. As such, `Contract.Require<T>()` will throw an `ArgumentException`.

This is not expected to occur, except for poorly designed custom exceptions. Regardless, this handling helps ensure that in those scenarios, clear guidance is provided.