Ensure custom exception type is explicitly thrown
As part of the C# 9.0 update, we were using the implicit `new()` for creating new exceptions. But there isn't enough context for the method to know that it should throw an exception of type `T` here! Fortunately, this scenario was only expected for scenarios where there was no `errorMessage` parameter which is rare, and thus probably didn't affect most consumers. (Even when just setting up a basic guard clause, we usually pass the name of the parameter as the message.)

This allows us to remove a legacy suppression regarding throwing an `Exception`.