Fixed test for `Contract.Assume<>()`
For reasons that aren't entirely clear, the `Contract.Assume<>()` test was specifying an `IndexOutOfRangeException`, but expecting an `InvalidOperationException`—and, apparently, passing! That should never have been setup that way, and it's unclear why it was working previously!

Regardless, this (correctly) fails now, _possibly_ due to a bug fix in throwing generic exceptions without a `errorMessage` (dc671719)—though that's unexpected since this test _does_ include an `errorMessage`.

While I'm baffled by this peculiar behavior, I'm not going to dive into it since a) it's confirmed to be working as expected now, b) all other unit tests relying on `Contract.Assume<>` (directly or indirectly) are working, and c) we haven't had any bug reports about this.