Introduce test for `GetHierarchicalRoot()` exception handling
This verifies that if the `navigationRoot` cannot be found based on the current `Topic`, then the `defaultRoot` is used—and, if that can't be found, an `if that isn't configured, then an `ArgumentOutOfRangeException` is thrown. Alternatively, if the `defaultRoot` is missing, then an `ArgumentNullException` is thrown. The latter isn't generally expected, but it's still possible for implementers to override the default value with e.g. an empty string.