Removed condition to allow nested topics with `[Flatten]`
There is currently no attribute that enables this condition, nor any scenario where this condition would be enabled. Perhaps there should be! But, if so, that will require further consideration and likely the introduction of a new attribute—or, at least, additional options on an existing attribute—to support. If we want to support that, we'll want to address that as part of a feature request. In the meanwhile, it should be removed to avoid confusion.