Introduce test to ensure cached mappings are unique per type
The `CachedTopicMappingService` distinguishes the cache based on `Type` to ensure that the same topic mapped as two different types are not confused with one another. This was not validated via unit tests, however, since the test for `CachedTopicMappingService` only called `MapAsync()` and not `MapAsync<T>()` which includes the type. This is remedied.