Introduced test for `ResetAttributeDescriptors()`
The `AttributeDescriptors` property is dynamically evaluated on the first access and then the results are cached. If an `AttributeDescriptor` is added or removed from the current topic or any of its ascendants then the `AttributeDescriptors` collection needs to be reset so the values can be reevaluated. This can be done using the `ResetAttributeDescriptors()` method which resets the `AttributeDescriptor` collection not only on the target `Topic` but also all of its descendants. This test validates that this works correctly.