Excluded `AttributeDescriptor` properties from code coverage
While these contain basic logic, it is standard logic that acts as a pass-through to the (well-validated) `AttributeCollection.SetValue()` method, so there's no need to test them individually.

Since the `AttributeDescriptor` is entirely made up of these properties, the class could be marked as `[ExcludeFromCodeCoverage]`, but I don't want to inadvertently exclude future methods should we introduce them.

Technically, almost all of these properties end up getting covered anyway by other tests. The one exception to this is `DisplayGroup`. But that's somewhat incidental and if those tests are modified, it shouldn't be expected that these properties will be explicitly tested.