Bug Fix: Empty `ReadOnlyKeyedTopicCollection<T>`
The introduction of a unit test for the empty `ReadOnlyKeyedTopicCollection` constructor (13ecc383) exposed that this use case was actually failing due to an errant `Contract.Requires()` guard clause in the underlying `ReadOnlyKeyedTopicCollection<T>`. This isn't necessary since the next line defines a default as a fallback in case the value is `null`.