Improve `listType` definition to support non-generics
Updated the `listType` definition to support non-generic types that implement generic interfaces by looping through available interfaces to identify any that implement `IList<>` and taking the last parameter type. This is the same logic that was previously updated in `TopicMappingService`, and is thus more consistent. As part of this, I was able to make the `listType` non-nullable, thus reducing the defensive logic later on.