Introduced test for skipping null `IAssociatedTopicBindingModel.UniqueKey`
Currently, the `ReverseTopicMappingService` will skip `IAssociatedTopicBindingModel`'s with a null or empty `UniqueKey`. That's an unintuitive and unnecessary limitation, which should be addressed in a future version (#42). Until then, the current behavior should be validated by a unit test—which will then need to be updated to test that this correctly overwrites existing topic references once that behavior is updated.