Introduce test for `TopicFactory.Create()` fallback for `AttributeDescriptor`
The `TopicFactory.Create()` method has a special condition for attribute descriptors to account for the fact that they need to be strongly typed as `AttributeDescriptor` instances to work correctly with the `ITopicRepository`, but the specific derived type may be in an external plugin and not always available (e.g., on applications not implementing the OnTopic Editor). In this case, any topic whose `ContentType` ends with `AttributeDescriptor` will fall back to `AttributeDescriptor` if a more specific type cannot be located in the `ITypeLookupService`. This test validates that behavior.