Introduce test to verify `IsDirty()` bypass if core attributes haven't changed
If the core attributes—i.e., `Key`, `ContentType`, and `Parent`—are set to their existing values then the `Topic` shouldn't be marked `IsDirty()` if it wasn't already. This test validates this behavior.

Note: This test currently fails due to a bug with `Parent` always tripping `IsDirty()`. This will be fixed in a subsequent commit.