Bug Fix: Existing topic with parent triggering `IsDirty()`
When constructing an existing topic—i.e., a `Topic` with an `Id`—the returned instance should not be marked `IsDirty()`. This is fine if the constructor only has a `Key`, `ContentType`, and `Id`, but fails if it _also_ has a `Parent`. To fix this, the topic is explicitly being `MarkClean()`ed at the end of the constructor, assuming the `Id` is set. This fixes a bug exposed by a new unit test (c0955abf).