Introduce test to validate `Topic.Key` change reindexes `Topic.Parent.Children`
When a `Topic.Key` is changed, the `Topic.Parent`'s `Topic.Children` collection won't automatically be reindexed to use the new `Topic.Key`. Instead, we have code in the `Key` setter that handle this. This unit test validates that this behavior works as expected.

Note: This doesn't handle other instances of `KeyedTopicCollection` that might be referencing the topic. There's no easy way to handle that.