Introduced test to ensure `GetValue()` treats empty as null
`TrackedRecord<T>.Value` should treat empty as `null`. That's true if the value is coming from a persistence store (which may not support nullable types) or, perhaps more commonly, form posts, where null values are treated as empty. Regardless of the source, `TrackedRecordCollection<T>.GetValue()` should paper over this difference by returning `null` if the `TrackedRecord<T>.Value` is empty. This test validates that is true.