Bug Fix: Remove previous reciprocal relationship with `SetValue()`
If `Topic.References.SetValue()` updates an existing topic reference, then the original reference's `IncomingRelationships` should be updated to remove that value, and the new reference's `IncomingRelationships` should be updated to add that value. A test introduced in the previous commit (74e74182) that demonstrated that this was not, in fact, working as expected. This is now fixed.