Updated test to evaluate parental inheritance as well as base inheritance
The original test ensures that a value was inherited from `Topic.BaseTopic`, but not that it was inherited from `Topic.Parent`. This corrects that by combining both conditions into a single test. (This is effectively tested by `AttributeCollectionTest` as well, so we're not too worried about breaking this up into more granular parts.)