Bug Fix: `Topic.References.Clear()` removes `IncomingRelationships`
When `Topic.References.Clear()` is called, the `Topic` should be removed from all referenced topics' `IncomingRelationships` collection. This logic was already in place, but it happened _after_ the current collection had already been cleared, and thus the loop over each reference failed to return any values. Whoops!

This was exposed by an update to an existing unit test to validate this behavior (28f064d8).