Update select tests to use `CachedTopicRepository`
The `StubTopicRepository` that's used for the `_topicRepository` reduces unnecessary intermediaries, but it is intended to support tests, not to be tested. As such, items that require actual testing of in-memory logic should go through the `CachedTopicRepository` instead.

(We may want to go back to running most items through the `CachedTopicRepository`, but that's something I need to put some more thought into.)