Call methods twice to check null conditions
When removing, clearing, or setting a null item, call the `Remove()`, `Clear()`, or `SetValue()` twice. The first to ensure that `IncomingRelationships` is set—the main purpose of the test. The second to ensure that the same call to `IncomingRelationships` doesn't fail if `TrackedTopic<T>.Value` is now null. This is a bit sloppy, but helps cover some of the rare but possible code paths covered by null checks.